Refactorisation IQ3D
TT est dans le titre
