+++ src/etat_descripteurs.py
@@ -20,7 +20,7 @@ from helpers.graph_tools import (
 )
 from helpers.tools_file import CheckConf
 
-from iq3d_descripteurs import DescripteurAnalyzer, get_configured_descriptors, graphe_desc_section
+from helpers.iq3d_descripteurs import DescripteurAnalyzer, get_configured_descriptors, graphe_desc_section
 
 
 def main(
